Separated into a module, exported macros
